fix: fix build error
