+++ src/app/(after-login)/course/[course-id]/_components/detail-info/detail-info.tsx
@@ -1,4 +1,5 @@
 import classNames from 'classnames/bind';
+
 import styles from './detail-info.module.scss';
 
 const cx = classNames.bind(styles);
