+++ src/app/(after-login)/course/_components/filtered-course/filtered-course.module.scss
@@ -1,0 +1,17 @@
+.filters{
+  display: flex;
+  gap: 0.5625rem;
+  padding: 0.5rem 1.25rem;
+}
+
+.course-list{
+  flex-grow: 1;
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+  padding: 0.5rem 1.25rem;
+  overflow-y: scroll;
+  &::-webkit-scrollbar {
+    display: none;
+  }
+}
