+++ src/app/(after-login)/course/page.module.scss
@@ -15,21 +15,3 @@
   padding: 0.375rem 1.25rem; 
   justify-content: center;
 }
-
-.filters{
-  display: flex;
-  gap: 0.5625rem;
-  padding: 0.5rem 1.25rem;
-}
-
-.course-list{
-  flex-grow: 1;
-  display: flex;
-  flex-direction: column;
-  gap: 1.5rem;
-  padding: 0.5rem 1.25rem;
-  overflow-y: scroll;
-  &::-webkit-scrollbar {
-    display: none;
-  }
-}
