+++ src/components/common/divider/index.tsx
@@ -1,4 +1,5 @@
 import classNames from 'classnames/bind';
+
 import styles from './divider.module.scss';
 
 const cx = classNames.bind(styles);
